+++ Cooperativa/Model/CodigosPostalesLocalidades.cs
@@ -8,7 +8,7 @@ namespace Model {
     public class CodigosPostalesLocalidades {
         public CodigosPostalesLocalidades() {
         }
-        public virtual int CplNumero { get; set; }
+        public virtual long CplNumero { get; set; }
         public virtual string CplDescripcion { get; set; }
         public virtual string CplCodigoPostal { get; set; }
         public virtual int LocNumero { get; set; }
